Updata dom with right position --save work
Dom diff is a bad idea, once you can trace all the change of your data,
and all the dom binded with your data, we never need a diff. Immu data
will help us when need to insert / modify a dom.

Also indroduce a DomCollection data structure to trace the dom tree, it
is a  tree that each node keep two pointers, they point to a slice of a
dom children first element and last element, with this data structure we
can always easy to insert the dom node in the right pos very quickly.